public/python/src/Proyectos/captura de objetos/DroneVision-YOLOv8/requirements.txt
# DroneVision-YOLOv8 - Dependencias del Sistema
# Plataforma de Análisis Aéreo Multi-Dominio

# Framework principal
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Procesamiento de imágenes y videos
opencv-python>=4.8.0
Pillow>=9.5.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Procesamiento de datos
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualización y gráficos
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilidades de sistema
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
psutil>=5.9.0

# Procesamiento de archivos
pathlib2>=2.3.7
glob2>=0.7

# Análisis de datos
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Utilidades de desarrollo (opcional)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Herramientas de logging
loguru>=0.7.0

# Utilidades de configuración
omegaconf>=2.3.0
hydra-core>=1.3.0

# Procesamiento de videos avanzado
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Utilidades de red (para descarga de modelos)
urllib3>=2.0.0
certifi>=2023.7.22

# Procesamiento de JSON
ujson>=5.8.0

# Utilidades de sistema operativo
platform>=1.0.8

# Dependencias específicas para YOLOv8
thop>=0.1.1
timm>=0.9.0
tensorboard>=2.13.0

# Dependencias para tracking
lap>=0.4.0
filterpy>=1.4.5

# Dependencias para segmentación
albumentations>=1.3.0
imgaug>=0.4.0

# Utilidades de métricas
pycocotools>=2.0.6
pycocotools-windows>=2.0.6; platform_system=="Windows"

# Dependencias para visualización avanzada
bokeh>=3.2.0
dash>=2.11.0
streamlit>=1.25.0

# Utilidades de configuración de GPU
nvidia-ml-py3>=7.352.0
pynvml>=11.5.0

# Dependencias para procesamiento de audio (si es necesario)
librosa>=0.10.0
soundfile>=0.12.0

# Utilidades de compresión
zipfile38>=0.0.3
tarfile>=0.0.1

# Dependencias para testing (opcional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Utilidades de documentación (opcional)
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Dependencias para deployment (opcional)
gunicorn>=21.2.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilidades de monitoreo (opcional)
wandb>=0.15.0
tensorboard>=2.13.0
mlflow>=2.5.0

# Dependencias específicas para Windows
pywin32>=306; platform_system=="Windows"
wmi>=1.5.1; platform_system=="Windows"

# Dependencias específicas para Linux
python-xlib>=0.33; platform_system=="Linux"

# Dependencias específicas para macOS
pyobjc>=9.0; platform_system=="Darwin"

# Utilidades de seguridad
cryptography>=41.0.0
keyring>=24.2.0

# Dependencias para procesamiento de archivos grandes
h5py>=3.9.0
zarr>=2.16.0

# Utilidades de configuración avanzada
click>=8.1.0
rich>=13.5.0
typer>=0.9.0

# Dependencias para optimización
numba>=0.57.0
cython>=3.0.0

# Utilidades de logging avanzado
structlog>=23.1.0
colorlog>=6.7.0

# Dependencias para procesamiento de datos geográficos (opcional)
geopandas>=0.13.0
folium>=0.14.0
